ai/backend/requirements.txt
fastapi
uvicorn
pytesseract
pdf2image
pdfplumber
pillow
opencv-python
numpy
pandas
python-dateutil
dateparser
spacy
